PiteYela/requirements.txt
PyQt6>=6.6.0
pyinstaller>=6.0.0
bcrypt>=4.0.0
python-escpos>=3.0.0
reportlab>=4.0.0
openpyxl>=3.1.0

